+++ http4k-format/jackson-yaml/src/main/kotlin/org/http4k/format/ConfigurableJacksonYaml.kt
@@ -5,6 +5,7 @@ import org.http4k.core.Body
 import org.http4k.core.ContentType
 import org.http4k.core.ContentType.Companion.TEXT_YAML
 import org.http4k.core.HttpMessage
+import org.http4k.core.with
 import org.http4k.lens.BiDiBodyLensSpec
 import org.http4k.lens.ContentNegotiation
 import org.http4k.lens.ContentNegotiation.Companion.None
@@ -33,6 +34,8 @@ open class ConfigurableJacksonYaml(val mapper: ObjectMapper, override val defaul
         contentNegotiation: ContentNegotiation = None
     ): BiDiBodyLensSpec<T> =
         httpBodyLens(description, contentNegotiation, defaultContentType).map(mapper.read(), mapper.write())
+
+    inline fun <reified T: Any, R: HttpMessage> R.with(t: T): R = with<R>(Body.auto<T>().toLens() of t)
 }
 
 inline operator fun <reified T : Any> ConfigurableJacksonYaml.invoke(msg: HttpMessage): T = autoBody<T>().toLens()(msg)